requirements.txt
joblib==1.0.1
scikit-learn==0.22.2.post1
sklearn-pandas==1.8.0
streamlit==0.80.0
pyngrok==4.1.1
ipykernel==5.5.3
matplotlib==3.2.2
numpy==1.19.5
pandas==1.1.5
virtualenv==16.0.0
